Refactor folders for better visibility of fragments and their logic
