+++ starter/app/src/main/java/com/udacity/shoestore/Login/login_destination.kt
@@ -1,25 +1,24 @@
-package com.udacity.shoestore
+package com.udacity.shoestore.Login
 
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.udacity.shoestore.R
 
-// TODO: Rename parameter arguments, choose names that match
-// the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
-private const val ARG_PARAM1 = "param1"
-private const val ARG_PARAM2 = "param2"
+const val ARG_PARAM1 = "param1"
+const val ARG_PARAM2 = "param2"
 
 /**
  * A simple [Fragment] subclass.
- * Use the [welcome_destination.newInstance] factory method to
+ * Use the [login_destination.newInstance] factory method to
  * create an instance of this fragment.
  */
-class welcome_destination : Fragment() {
+class login_destination : Fragment() {
     // TODO: Rename and change types of parameters
-    private var param1: String? = null
-    private var param2: String? = null
+    var param1: String? = null
+    var param2: String? = null
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -34,7 +33,7 @@ class welcome_destination : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_welcome_destination, container, false)
+        return inflater.inflate(R.layout.fragment_login_destination, container, false)
     }
 
     companion object {
@@ -44,12 +43,12 @@ class welcome_destination : Fragment() {
          *
          * @param param1 Parameter 1.
          * @param param2 Parameter 2.
-         * @return A new instance of fragment welcome_destination.
+         * @return A new instance of fragment login_destination.
          */
         // TODO: Rename and change types and number of parameters
         @JvmStatic
         fun newInstance(param1: String, param2: String) =
-            welcome_destination().apply {
+            login_destination().apply {
                 arguments = Bundle().apply {
                     putString(ARG_PARAM1, param1)
                     putString(ARG_PARAM2, param2)
