+++ starter/app/src/main/java/com/udacity/shoestore/models/ShoeViewModel.kt
@@ -1,0 +1,7 @@
+package com.udacity.shoestore.models
+
+import androidx.lifecycle.ViewModel
+
+class ShoeViewModel : ViewModel() {
+
+}
