+++ starter/app/src/main/res/layout/fragment_instruction_destination.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Instruction_destination">
+    tools:context=".Instructions.Instruction_destination">
 
     <LinearLayout
         android:layout_width="match_parent"
