+++ starter/app/src/main/res/navigation/main_navigation.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/login_destination"
-        android:name="com.udacity.shoestore.LoginFragment"
+        android:name="com.udacity.shoestore.Login.LoginFragment"
         android:label="fragment_login_destination"
         tools:layout="@layout/fragment_login_destination" >
         <action
@@ -16,7 +16,7 @@
     </fragment>
     <fragment
         android:id="@+id/welcome_destination"
-        android:name="com.udacity.shoestore.WelcomeFragment"
+        android:name="com.udacity.shoestore.Welcome.WelcomeFragment"
         android:label="fragment_welcome_destination"
         tools:layout="@layout/fragment_welcome_destination" >
         <action
@@ -25,7 +25,7 @@
     </fragment>
     <fragment
         android:id="@+id/instruction_destination"
-        android:name="com.udacity.shoestore.Instruction_destination"
+        android:name="com.udacity.shoestore.Instructions.Instruction_destination"
         android:label="fragment_instruction_destination"
         tools:layout="@layout/fragment_instruction_destination" />
 </navigation>
